app/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dataclasses-json==0.6.2
distro==1.8.0
exceptiongroup==1.1.3
filelock==3.13.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.335
langchain-experimental==0.0.40
langsmith==0.0.64
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.2.4
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.25.0
pyarrow==14.0.1
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.50
streamlit==1.28.1
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.2
distro==1.8.0
exceptiongroup==1.1.3
filelock==3.13.1
fonttools==4.44.3
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.335
langchain-experimental==0.0.40
langsmith==0.0.64
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.2.4
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.18.0
protobuf==4.25.0
pyarrow==14.0.1
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.50
streamlit==1.28.1
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0
